removing the unnecessary data structures

+++ code/mmim/algorithms.hpp
@@ -6,11 +6,11 @@
 #include <vector>
 #include "graph.hpp"
 
-std::vector<float> random(int init, int rep, int k, int gap, Graph &graph);
-std::vector<float> max_deg(int init, int rep, int k, int gap, Graph &graph);
-std::vector<float> max_deg_adv(int init, int rep, int k, int gap, Graph &graph);
-std::vector<float> min_deg(int init, int rep, int k, int gap, Graph &graph);
-std::vector<float> min_deg_adv(int init, int rep, int k, int gap, Graph &graph);
-std::vector<float> k_gonz(int init, int rep, int k, int gap, Graph graph);
+std::vector<int> random(int init, int rep, int k, int gap, Graph &graph);
+std::vector<int> max_deg(int init, int rep, int k, int gap, Graph &graph);
+std::vector<int> max_deg_adv(int init, int rep, int k, int gap, Graph &graph);
+std::vector<int> min_deg(int init, int rep, int k, int gap, Graph &graph);
+std::vector<int> min_deg_adv(int init, int rep, int k, int gap, Graph &graph);
+std::vector<int> k_gonz(int init, int rep, int k, int gap, Graph graph);
 
 #endif // ALGORITHMS_HPP
